// enums/leaderEducationalLevel.enum.ts
export enum ELeaderEducationalLevel {
  SEM_ESCOLARIDADE = "Sem escolaridade",
  FUNDAMENTAL_INCOMPLETO = "Fundamental Incompleto",
  FUNDAMENTAL_COMPLETO = "Fundamental Completo",
  SEGUNDO_GRAU_INCOMPLETO = "Segundo Grau Incompleto",
  SEGUNDO_GRAU_COMPLETO = "Segundo Grau Completo",
  SUPERIOR_INCOMPLETO = "Superior Incompleto",
  SUPERIOR_COMPLETO = "Superior Completo",
  POS_GRADUADO = "Pós-graduado",
}