// enums/leaderSocialProgram.enum.ts
export enum ELeaderSocialProgram {
  AUXILIO_GAS = "Auxílio Gás",
  BOLSA_FAMILIA = "Bolsa Família",
  BPC = "Benefício de Prestação Continuada",
  BRASIL_SORRIDENTE = "Brasil Sorridente",
  FARMACIA_POPULAR = "Farmácia Popular",
  MINHA_CASA_MINHA_VIDA = "Minha Casa, Minha Vida",
  PAA = "Programa de Aquisição de Alimentos",
  PRONATEC = "Programa Nacional de Acesso ao Ensino Técnico e Emprego",
  TARIFA_SOCIAL_ENERGIA = "Tarifa Social de Energia Elétrica",
}